pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rasc"
version = "0.1.0"
description = "Companion codebase for Retrieve, Then Classify: Corpus-Grounded Automation of Clinical Value Set Authoring."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "accelerate>=0.30",
  "faiss-cpu>=1.8.0",
  "joblib>=1.3",
  "lightgbm>=4.3",
  "numpy>=1.26",
  "requests>=2.31",
  "scikit-learn>=1.4",
  "sentence-transformers>=2.7",
  "torch>=2.2",
  "tqdm>=4.66",
  "transformers>=4.41",
]

[tool.setuptools]
py-modules = []
